% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematica2r.R
\name{mathematica2r}
\alias{mathematica2r}
\title{Conversion from Mathematica to R}
\usage{
mathematica2r(data, numerics, digits = 4)
}
\arguments{
\item{data}{The dataset}

\item{numerics}{What columns are to be considered numeric in R}

\item{digits}{To what digit to round numerical values?}
}
\description{
Few adjustments to the data generated by Mathematica, for easy handling in R.
}
