% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_divergence_map.R
\name{plot_divergence_map}
\alias{plot_divergence_map}
\title{Plot divergence map}
\usage{
plot_divergence_map(data, xstart, keep_legend = TRUE)
}
\arguments{
\item{data}{A dataset of singularities for the one-patch model}

\item{xstart}{The initial trait value}

\item{keep_legend}{Whether to return the legend}
}
\description{
Heatmap of trait divergence between the two patches in the absence of dispersal (one-patch model). A transparent extra layer indicates whether branching occurs within both patch (relevant only if both patches are mirror images of each other).
}
