% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_adaptive_dynamics.R
\name{predict_adaptive_dynamics}
\alias{predict_adaptive_dynamics}
\title{Predict the adaptive dynamics}
\usage{
predict_adaptive_dynamics(data, xstart)
}
\arguments{
\item{data}{A dataset of singularities}

\item{xstart}{The initial trait value}
}
\description{
Makes a summary dataset containing the attained singularity given a starting trait value and a map of singular points throughout parameter space.
}
